updated the look of the planet table... ready for final testing and timing
git-svn-id: http://phenomena-client.googlecode.com/svn/trunk/HelioroomTeacherDashboard@61 b2d1b222-19ed-11df-8faf-b330578705e8

+++ src/ltg/phenomena/helioroom/Planet.java
@@ -20,7 +20,7 @@ public class Planet {
 	private int secToNextWin = -2;
 	
 	
-	public Planet(String name, String color, String colorName, int classOrbitalTime, int startPosition) {
+	public Planet(String name, String color, String colorName, int classOrbitalTime, double startPosition) {
 		this.name = name;
 		this.color = color.replaceAll("0x", "#");
 		this.colorName = colorName;
